refactor: split pinstore
